requirements.txt
redis
matplotlib
numpy
torch
fastapi
pydantic
pillow
prometheus_client
uvicorn

